modified controllers architecture added redirect from base link to notes
