+++ src/main/java/me/savcheg/notebookwebspring/controllers/MainViewController.java
@@ -3,18 +3,13 @@ package me.savcheg.notebookwebspring.controllers;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.servlet.view.RedirectView;
 
 @Controller
-@RequestMapping("/notes")
-public class ViewController {
-
+@RequestMapping("/")
+public class MainViewController {
     @GetMapping
-    public String getAllNotesPage() {
-        return "notes";
-    }
-
-    @GetMapping("/new")
-    public String getNewNotePage() {
-        return "new_note";
+    public RedirectView redirectToNotes() {
+        return new RedirectView("/notes");
     }
 }
